requirements.txt
asgiref==3.7.2
Django==4.0.0
django-tinymce==3.6.1
Pillow==10.0.0
PyMySQL==1.1.0
sqlparse==0.4.4
